+++ A1/project/package.json
@@ -8,7 +8,8 @@
   },
   "dependencies": {
     "react": "16.2.0",
-    "react-native": "0.52.1"
+    "react-native": "0.52.1",
+    "react-native-router-flux": "^4.0.0-beta.28"
   },
   "devDependencies": {
     "babel-jest": "22.1.0",
